Product model modifications
Add attributes to Product: Quantity, description_1 and description_2

Improve the quality of the attributes. The descriptions and titles have
more meaning

Fix the layout from the scrollable product div in home, category index
and product show view.

Style product show view

The images from the products are now a link to the respective product show
action