Cart Products join table replaced by a model
Create CartItem model. An id and a new quantity attribute were the reasons
for this.

Set up relationships between Cart, Product and CartItem.

Create form on product#show view for the quantity

Fix the rails helper on Cart #show view to render the images and
fix the delete buttons

The change from the join table to a full model is complete